requirements.txt
streamlit==1.18.1
langchain==0.0.200
openai==0.27.0
PyYAML==6.0
emoji